Pesquisar Contato - nao traz dados do banco =/

+++ testeswicket/src/main/java/br/com/testeswicket/contato/Contato.java
@@ -5,16 +5,17 @@ import java.io.Serializable;
 public class Contato implements Serializable{
 	private static final long serialVersionUID = 6982487603929829074L;
 	
-	private Long id;
+	private Integer id;
 	private String nome;
 	private String email;
 	private String telefone;
 	
 	
-	public Long getId() {
+
+	public Integer getId() {
 		return id;
 	}
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 	public String getNome() {
